backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7

# File Handling
Pillow==10.0.1
django-storages==1.14.2
boto3==1.29.7

# API Documentation
drf-spectacular==0.26.5

# Utilities
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0

# Development
django-debug-toolbar==4.2.0